--- apm-agent-plugins/apm-apache-httpclient-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,4 +1,0 @@
-co.elastic.apm.agent.httpclient.ApacheHttpClientInstrumentation
-co.elastic.apm.agent.httpclient.ApacheHttpAsyncClientInstrumentation
-co.elastic.apm.agent.httpclient.ApacheHttpAsyncClientRedirectInstrumentation
-co.elastic.apm.agent.httpclient.LegacyApacheHttpClientInstrumentation
