+++ apm-agent-plugins/apm-apache-httpclient/apm-apache-httpclient3-plugin/src/main/java/co/elastic/apm/agent/httpclient/v3/HttpClient3RequestHeaderAccessor.java
@@ -1,0 +1,53 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package co.elastic.apm.agent.httpclient.v3;
+
+import co.elastic.apm.agent.impl.transaction.TextHeaderGetter;
+import co.elastic.apm.agent.impl.transaction.TextHeaderSetter;
+import org.apache.commons.httpclient.Header;
+import org.apache.commons.httpclient.HttpMethod;
+
+import javax.annotation.Nullable;
+
+public class HttpClient3RequestHeaderAccessor implements TextHeaderGetter<HttpMethod>, TextHeaderSetter<HttpMethod> {
+
+    public static HttpClient3RequestHeaderAccessor INSTANCE = new HttpClient3RequestHeaderAccessor();
+
+    @Nullable
+    @Override
+    public String getFirstHeader(String headerName, HttpMethod carrier) {
+        Header[] headers = carrier.getRequestHeaders(headerName);
+        return headers != null && headers.length > 0 ? headers[0].getValue() : null;
+    }
+
+    @Override
+    public <S> void forEach(String headerName, HttpMethod carrier, S state, HeaderConsumer<String, S> consumer) {
+        Header[] headers = carrier.getRequestHeaders();
+        if (headers != null) {
+            for (Header header : headers) {
+                consumer.accept(header.getValue(), state);
+            }
+        }
+    }
+
+    @Override
+    public void setHeader(String headerName, String headerValue, HttpMethod carrier) {
+        carrier.setRequestHeader(headerName, headerValue);
+    }
+}
