+++ apm-agent-plugins/apm-apache-httpclient/apm-apache-httpclient3-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,0 +1,1 @@
+co.elastic.apm.agent.httpclient.v3.HttpClient3Instrumentation
