+++ apm-agent-plugins/apm-apache-httpclient/apm-apache-httpclient3-plugin/src/test/java/co/elastic/apm/agent/httpclient/v3/HttpClient3VersionIT.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package co.elastic.apm.agent.httpclient.v3;
+
+import co.elastic.apm.agent.TestClassWithDependencyRunner;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Arrays;
+import java.util.List;
+
+@RunWith(Parameterized.class)
+public class HttpClient3VersionIT {
+
+    private final TestClassWithDependencyRunner runner;
+
+    public HttpClient3VersionIT(String dependency) throws Exception {
+        this.runner = new TestClassWithDependencyRunner(List.of(dependency), HttpClient3InstrumentationTest.class);
+    }
+
+    @Parameterized.Parameters(name = "{0}")
+    public static Iterable<Object[]> data() {
+        return Arrays.asList(new Object[][]{
+            {"commons-httpclient:commons-httpclient:3.0"},
+            {"commons-httpclient:commons-httpclient:3.1"}
+        });
+    }
+
+    @Test
+    public void testVersion() {
+        runner.run();
+    }
+}
