+++ apm-agent-plugins/apm-apache-httpclient/apm-apache-httpclient4-plugin/pom.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <artifactId>apm-apache-httpclient</artifactId>
+        <groupId>co.elastic.apm</groupId>
+        <version>1.34.2-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>apm-apache-httpclient4-plugin</artifactId>
+    <name>${project.groupId}:${project.artifactId}</name>
+
+    <properties>
+        <apm-agent-parent.base.dir>${project.basedir}/../../..</apm-agent-parent.base.dir>
+    </properties>
+
+    <dependencies>
+        <!--
+        this dependency duplicates the transitive one we get from 'httpasyncclient'
+        but keeping it explicit avoids relying on transitive dependency
+        -->
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+            <version>4.5.6</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpasyncclient</artifactId>
+            <version>4.1.4</version>
+            <scope>provided</scope>
+        </dependency>
+
+    </dependencies>
+
+</project>
