+++ apm-agent-plugins/apm-apache-httpclient/apm-apache-httpclient4-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,0 +1,4 @@
+co.elastic.apm.agent.httpclient.v4.ApacheHttpClientInstrumentation
+co.elastic.apm.agent.httpclient.v4.ApacheHttpAsyncClientInstrumentation
+co.elastic.apm.agent.httpclient.v4.ApacheHttpAsyncClientRedirectInstrumentation
+co.elastic.apm.agent.httpclient.v4.LegacyApacheHttpClientInstrumentation
