+++ apm-agent-plugins/pom.xml
@@ -27,7 +27,7 @@
         <module>apm-spring-webflux</module>
         <module>apm-reactor-plugin</module>
         <module>apm-api-plugin</module>
-        <module>apm-apache-httpclient-plugin</module>
+        <module>apm-apache-httpclient</module>
         <module>apm-spring-resttemplate</module>
         <module>apm-httpclient-core</module>
         <module>apm-httpserver-core</module>
